+++ js/functions.js
@@ -22,7 +22,7 @@ const extractNumbers = (string) => {
   let result = '';
   string = string.toString();
   for (let i = 0; i <= string.length - 1; i++) {
-    if (Number.isNaN (parseInt (string[i], 10)) === false) {
+    if (Number.isNaN(parseInt(string[i], 10)) === false) {
       result += string[i];
     }
   }
